Refactor legend control implementation
