+++ client/src/components/map/Legend/Legend.css
@@ -1,0 +1,28 @@
+.legend-button {
+    width: 40px;
+    height: 40px;
+    background-color: #121212;
+    border: none;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    padding: 0;
+    border-radius: 0;
+}
+
+.legend-icon {
+    font-size: 20px;
+    color: white;
+}
+
+.watershed-left {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.watershed-right {
+    display: flex;
+    gap: 10px;
+}
