+++ client/src/components/map/Legend/Legend.tsx
@@ -1,0 +1,69 @@
+import { useState } from 'react';
+import { FaListUl, FaXmark, FaEye, FaExpand } from "react-icons/fa6";
+import './Legend.css';
+
+export default function LegendControl() {
+  const [isLegendOpen, setIsLegendOpen] = useState(false);
+
+  const toggleLegend = () => setIsLegendOpen((prev) => !prev);
+
+  return (
+    <div className="leaflet-bar leaflet-control">
+      <button
+        onClick={toggleLegend}
+        className="legend-button"
+        aria-label="Toggle legend"
+        title="Toggle legend"
+      >
+        {isLegendOpen ? (
+          <FaXmark className="legend-icon" />
+        ) : (
+          <FaListUl className="legend-icon" />
+        )}
+      </button>
+
+      {isLegendOpen && (
+        <div className="legend-modal">
+          <div className="legend-content">
+            <div className="watershed-container">
+              <div className="watershed-left">
+                <div style={{ background: '#00FF7F', width: 20, height: 20, display: 'inline-block', borderRadius: 5 }}></div>
+                <span className="legend-text" style={{ fontWeight: 'bold' }}>Tier 1 watersheds</span>
+              </div>
+              <div className="watershed-right">
+                <FaExpand
+                  className="legend-icon"
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => alert('Show only icon clicked')}
+                />
+                <FaEye
+                  className="legend-icon"
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => alert('Hide icon clicked')}
+                />
+              </div>
+            </div>
+            <div className="watershed-container">
+              <div className="watershed-left">
+                <div style={{ background: '#0000FF', width: 20, height: 20, display: 'inline-block', borderRadius: 5 }}></div>
+                <span className="legend-text" style={{ fontWeight: 'bold' }}>Tier 2 watersheds</span>
+              </div>
+              <div className="watershed-right">
+                <FaExpand
+                  className="legend-icon"
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => alert('Show only icon clicked')}
+                />
+                <FaEye
+                  className="legend-icon"
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => alert('Hide icon clicked')}
+                />
+              </div>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
